requirements.txt
langchain-core==0.2.25
langchain-text-splitters==0.2.2
langsmith==0.1.139
llama-cloud==0.1.4
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.21
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.13
nltk==3.9.1
numpy==1.26.4
openai==1.53.0
PyMuPDF==1.24.13
pymupdf4llm==0.0.17
pypdf==4.3.1
streamlit==1.26.0
tiktoken==0.8.0

